Performance improvement, only calculate distance once
Previous version took nearly a minute, this one only 4 seconds.

The old version calculated the closest point to each point in
the set of neighbors, recursively. This meant double-computing
all points on an edge, plus it had to constantly keep track of all
neighbor points on a queue and repeatedly cull duplicates.

This method only tracks in-bounds points. It could be smaller by
determining a smaller bound that excludes all exterior groups,
but I didn't want to do that math and 4 seconds isn't bad.