Problem 18/15 - Simulate a whole ass video game, I hate this one it's stupid
I fixed the bug though, the issue is to first find every possible *path* to every
target, choose the paths that are shortest, sort them by the reading order of the
targets, choose the target with the first reading order, then select from those paths
the one with the first step in the first reading order.

I was only finding that *a* path existed and it was minimal, not which target it
was going to. So if a target was in first reading order, but a step had to go in a
bad reading order (down for example to get around a barrier) that would be discarded
because of the step's reading order, even though the target is closer/higher reading
order